Add --global flag to use command

+++ toys/use.rb
@@ -2,12 +2,19 @@
 
 desc 'Switch a version to use'
 
-required_arg :version
+flag :global do
+  desc 'Set global version'
+  accept TrueClass
+  default false
+end
+
+required_arg :version do
+  desc 'Version to use'
+  accept(/^\d+(\.\d+){0,2}$/)
+end
 
 def run
-  if File.exist? GLOBAL_VERSION_FILE
-    File.write(LOCAL_VERSION_FILE, version)
-  else
-    File.write(GLOBAL_VERSION_FILE, version)
-  end
+  File.write(GLOBAL_VERSION_FILE, version) if global || (!File.exist? GLOBAL_VERSION_FILE)
+
+  File.write(LOCAL_VERSION_FILE, version) unless global
 end
